+++ dcpp/ThrottleManager.h
@@ -1,5 +1,5 @@
 /* 
- * Copyright (C) 2009-2010 Jacek Sieka, arnetheduck on gmail point com
+ * Copyright (C) 2009-2011 Jacek Sieka, arnetheduck on gmail point com
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -89,7 +89,7 @@ namespace dcpp
 		void waitToken();
 
 		// TimerManagerListener
-		void on(TimerManagerListener::Second, uint32_t /* aTick */) throw();
+		void on(TimerManagerListener::Second, uint64_t /* aTick */) throw();
 	};
 
 }	// namespace dcpp
